+++ tests/test_intersect_match.py
@@ -11,8 +11,6 @@ from __future__ import division
 from __future__ import print_function
 from __future__ import unicode_literals
 
-from eppy.iddcurrent import iddcurrent
-from geomeppy.eppy_patches import IDF
 from geomeppy.intersect_match import get_adjacencies
 from geomeppy.intersect_match import getidfsurfaces
 from geomeppy.intersect_match import intersect
@@ -21,6 +19,9 @@ from geomeppy.intersect_match import is_hole
 from geomeppy.intersect_match import match_idf_surfaces
 from geomeppy.intersect_match import unique
 from geomeppy.polygons import Polygon3D
+
+from eppy.iddcurrent import iddcurrent
+from geomeppy.eppy_patches import IDF
 from six import StringIO